requirements.txt
Django
django-leaflet
django-geojson
django-forms-bootstrap
pillow
psycopg2
